Various updates across the codebase:
- Move package into `src` subdirectory, which allows running import tests from
  the top of the source tree on the installed package.

- Support changes to the logging level with the FLACARRAY_LOGLEVEL environment
  variable.

- Test with libflac-1.5.0 and make this the default version vendored with our
  wheels.

- Support both the legacy Zarr-2 and the new Zarr-3 APIs

- Remove pinned numpy-2.0.x during build.  All numpy-2 versions at build time are
  forward and backward compatible with runtime numpy versions.

- Combine the deploy and wheel test workflows into a single workflow with
  deployment to PyPI conditional on a github release.
