+++ .github/workflows/wheels.yml
@@ -1,7 +1,10 @@
 
-name:  Test Binary Wheels
+name:  Binary Wheels
 
-on: workflow_dispatch
+on:
+  workflow_dispatch:
+  release:
+    types: [ published ]
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
@@ -109,5 +112,21 @@ jobs:
 
       - uses: actions/upload-artifact@v4
         with:
-          name: wheels_cp${{ matrix.python }}-${{ matrix.builder }}_${{ matrix.arch }}
+          name: flacarray-cp${{ matrix.python }}-${{ matrix.builder }}_${{ matrix.arch }}
           path: ./wheelhouse/flacarray*cp${{ matrix.python }}-${{ matrix.builder }}*${{ matrix.arch }}*.whl
+
+  upload_pypi:
+    needs: build_wheels
+    runs-on: ubuntu-latest
+    permissions:
+      id-token: write
+    if: github.event_name == 'release' && github.event.action == 'published'
+    steps:
+      - uses: actions/download-artifact@v4
+        with:
+          # Unpacks all artifacts into dist/
+          pattern: flacarray-*
+          path: dist
+          merge-multiple: true
+
+      - uses: pypa/gh-action-pypi-publish@release/v1
