+++ packaging/pip_build_requirements.txt
@@ -2,5 +2,6 @@ cmake
 cython
 meson
 meson-python
+numpy
 h5py
 zarr
