+++ src/flacarray/compress.py
