+++ src/flacarray/decompress.py
