+++ src/flacarray/tests/zarr.py
@@ -10,15 +10,14 @@ import numpy as np
 
 from ..array import FlacArray
 from ..demo import create_fake_data
-from ..hdf5 import write_array, read_array
-from ..hdf5_utils import H5File, have_hdf5
+from ..zarr import have_zarr, write_array, read_array, ZarrGroup
 from ..mpi import use_mpi, MPI
 
-if have_hdf5:
-    import h5py
+if have_zarr:
+    import zarr
 
 
-class HDF5Test(unittest.TestCase):
+class ZarrTest(unittest.TestCase):
     def setUp(self):
         fixture_name = os.path.splitext(os.path.basename(__file__))[0]
         if use_mpi:
@@ -27,8 +26,8 @@ class HDF5Test(unittest.TestCase):
             self.comm = None
 
     def test_direct_write_read(self):
-        if not have_hdf5:
-            print("h5py not available, skipping tests", flush=True)
+        if not have_zarr:
+            print("zarr not available, skipping tests", flush=True)
             return
         if self.comm is None:
             rank = 0
@@ -65,11 +64,11 @@ class HDF5Test(unittest.TestCase):
         if self.comm is not None:
             tmppath = self.comm.bcast(tmppath, root=0)
 
-        i32_file = os.path.join(tmppath, "data_i32.h5")
-        with H5File(i32_file, "w", comm=self.comm) as hf:
+        i32_file = os.path.join(tmppath, "data_i32.zarr")
+        with ZarrGroup(i32_file, mode="w", comm=self.comm) as zf:
             write_array(
                 input32,
-                hf.handle,
+                zf,
                 level=5,
                 quanta=None,
                 precision=None,
@@ -78,9 +77,9 @@ class HDF5Test(unittest.TestCase):
             )
         if self.comm is not None:
             self.comm.barrier()
-        with H5File(i32_file, "r", comm=self.comm) as hf:
+        with ZarrGroup(i32_file, mode="r", comm=self.comm) as zf:
             check_i32 = read_array(
-                hf.handle,
+                zf,
                 keep=None,
                 stream_slice=None,
                 keep_indices=False,
@@ -89,11 +88,11 @@ class HDF5Test(unittest.TestCase):
                 use_threads=True,
             )
 
-        i64_file = os.path.join(tmppath, "data_i64.h5")
-        with H5File(i64_file, "w", comm=self.comm) as hf:
+        i64_file = os.path.join(tmppath, "data_i64.zarr")
+        with ZarrGroup(i64_file, mode="w", comm=self.comm) as zf:
             write_array(
                 input64,
-                hf.handle,
+                zf,
                 level=5,
                 quanta=None,
                 precision=None,
@@ -102,9 +101,9 @@ class HDF5Test(unittest.TestCase):
             )
         if self.comm is not None:
             self.comm.barrier()
-        with H5File(i64_file, "r", comm=self.comm) as hf:
+        with ZarrGroup(i64_file, mode="r", comm=self.comm) as zf:
             check_i64 = read_array(
-                hf.handle,
+                zf,
                 keep=None,
                 stream_slice=None,
                 keep_indices=False,
@@ -113,11 +112,11 @@ class HDF5Test(unittest.TestCase):
                 use_threads=True,
             )
 
-        f32_file = os.path.join(tmppath, "data_f32.h5")
-        with H5File(f32_file, "w", comm=self.comm) as hf:
+        f32_file = os.path.join(tmppath, "data_f32.zarr")
+        with ZarrGroup(f32_file, mode="w", comm=self.comm) as zf:
             write_array(
                 inputf32,
-                hf.handle,
+                zf,
                 level=5,
                 quanta=None,
                 precision=None,
@@ -126,9 +125,9 @@ class HDF5Test(unittest.TestCase):
             )
         if self.comm is not None:
             self.comm.barrier()
-        with H5File(f32_file, "r", comm=self.comm) as hf:
+        with ZarrGroup(f32_file, mode="r", comm=self.comm) as zf:
             check_f32 = read_array(
-                hf.handle,
+                zf,
                 keep=None,
                 stream_slice=None,
                 keep_indices=False,
@@ -137,11 +136,11 @@ class HDF5Test(unittest.TestCase):
                 use_threads=True,
             )
 
-        f64_file = os.path.join(tmppath, "data_f64.h5")
-        with H5File(f64_file, "w", comm=self.comm) as hf:
+        f64_file = os.path.join(tmppath, "data_f64.zarr")
+        with ZarrGroup(f64_file, mode="w", comm=self.comm) as zf:
             write_array(
                 inputf64,
-                hf.handle,
+                zf,
                 level=5,
                 quanta=None,
                 precision=None,
@@ -150,9 +149,9 @@ class HDF5Test(unittest.TestCase):
             )
         if self.comm is not None:
             self.comm.barrier()
-        with H5File(f64_file, "r", comm=self.comm) as hf:
+        with ZarrGroup(f64_file, mode="r", comm=self.comm) as zf:
             check_f64 = read_array(
-                hf.handle,
+                zf,
                 keep=None,
                 stream_slice=None,
                 keep_indices=False,
@@ -161,36 +160,38 @@ class HDF5Test(unittest.TestCase):
                 use_threads=True,
             )
 
+        if self.comm is not None:
+            self.comm.barrier()
         del tmppath
         del tmpdir
 
         if not np.array_equal(check_i32, input32):
             print(f"check_i32 = {check_i32}", flush=True)
             print(f"input_i32 = {input32}", flush=True)
-            print("FAIL on i32 roundtrip to hdf5", flush=True)
+            print("FAIL on i32 roundtrip to zarr", flush=True)
             self.assertTrue(False)
 
         if not np.array_equal(check_i64, input64):
             print(f"check_i64 = {check_i64}", flush=True)
             print(f"input_i64 = {input64}", flush=True)
-            print("FAIL on i64 roundtrip to hdf5", flush=True)
+            print("FAIL on i64 roundtrip to zarr", flush=True)
             self.assertTrue(False)
 
         if not np.allclose(check_f32, inputf32, atol=1e-6):
             print(f"check_f32 = {check_f32}", flush=True)
             print(f"input_f32 = {inputf32}", flush=True)
-            print("FAIL on f32 roundtrip to hdf5", flush=True)
+            print("FAIL on f32 roundtrip to zarr", flush=True)
             self.assertTrue(False)
 
         if not np.allclose(check_f64, inputf64, atol=1e-6):
             print(f"check_f64 = {check_f64}", flush=True)
             print(f"input_f64 = {inputf64}", flush=True)
-            print("FAIL on f64 roundtrip to hdf5", flush=True)
+            print("FAIL on f64 roundtrip to zarr", flush=True)
             self.assertTrue(False)
 
     def test_array_write_read(self):
-        if not have_hdf5:
-            print("h5py not available, skipping tests", flush=True)
+        if not have_zarr:
+            print("zarr not available, skipping tests", flush=True)
             return
         if self.comm is None:
             rank = 0
@@ -216,96 +217,92 @@ class HDF5Test(unittest.TestCase):
         check_i64 = None
 
         inputf32 = create_fake_data(data_shape, 1.0).astype(np.float32)
-        flcarr_f32 = FlacArray.from_array(
-            inputf32, mpi_comm=self.comm, use_threads=True
-        )
+        flcarr_f32 = FlacArray.from_array(inputf32, mpi_comm=self.comm)
         check_f32 = None
 
         inputf64 = create_fake_data(data_shape, 1.0)
-        flcarr_f64 = FlacArray.from_array(
-            inputf64, mpi_comm=self.comm, use_threads=True
-        )
+        flcarr_f64 = FlacArray.from_array(inputf64, mpi_comm=self.comm, use_threads=True)
         check_f64 = None
 
-        tmpdir = None
-        tmppath = None
-        if rank == 0:
-            tmpdir = tempfile.TemporaryDirectory()
-            tmppath = tmpdir.name
-        if self.comm is not None:
-            tmppath = self.comm.bcast(tmppath, root=0)
+        # tmpdir = None
+        # tmppath = None
+        # if rank == 0:
+        #     tmpdir = tempfile.TemporaryDirectory()
+        #     tmppath = tmpdir.name
+        # if self.comm is not None:
+        #     tmppath = self.comm.bcast(tmppath, root=0)
 
-        i32_file = os.path.join(tmppath, "data_i32.h5")
-        with H5File(i32_file, "w", comm=self.comm) as hf:
-            flcarr_i32.write_hdf5(hf.handle)
-        if self.comm is not None:
-            self.comm.barrier()
-        with H5File(i32_file, "r", comm=self.comm) as hf:
-            check_i32 = FlacArray.read_hdf5(hf.handle, mpi_comm=self.comm)
+        # i32_file = os.path.join(tmppath, "data_i32.zarr")
+        # with ZarrGroup(i32_file, mode="w", comm=self.comm) as zf:
+        #     flcarr_i32.write_zarr(zf)
+        # if self.comm is not None:
+        #     self.comm.barrier()
+        # with ZarrGroup(i32_file, mode="r", comm=self.comm) as zf:
+        #     check_i32 = FlacArray.read_zarr(zf, mpi_comm=self.comm)
 
-        i64_file = os.path.join(tmppath, "data_i64.h5")
-        with H5File(i64_file, "w", comm=self.comm) as hf:
-            flcarr_i64.write_hdf5(hf.handle)
-        if self.comm is not None:
-            self.comm.barrier()
-        with H5File(i64_file, "r", comm=self.comm) as hf:
-            check_i64 = FlacArray.read_hdf5(hf.handle, mpi_comm=self.comm)
+        # i64_file = os.path.join(tmppath, "data_i64.zarr")
+        # with ZarrGroup(i64_file, mode="w", comm=self.comm) as zf:
+        #     flcarr_i64.write_zarr(zf)
+        # if self.comm is not None:
+        #     self.comm.barrier()
+        # with ZarrGroup(i64_file, mode="r", comm=self.comm) as zf:
+        #     check_i64 = FlacArray.read_zarr(zf, mpi_comm=self.comm)
 
-        f32_file = os.path.join(tmppath, "data_f32.h5")
-        with H5File(f32_file, "w", comm=self.comm) as hf:
-            flcarr_f32.write_hdf5(hf.handle)
-        if self.comm is not None:
-            self.comm.barrier()
-        with H5File(f32_file, "r", comm=self.comm) as hf:
-            check_f32 = FlacArray.read_hdf5(hf.handle, mpi_comm=self.comm)
+        # f32_file = os.path.join(tmppath, "data_f32.zarr")
+        # with ZarrGroup(f32_file, mode="w", comm=self.comm) as zf:
+        #     flcarr_f32.write_zarr(zf)
+        # if self.comm is not None:
+        #     self.comm.barrier()
+        # with ZarrGroup(f32_file, mode="r", comm=self.comm) as zf:
+        #     check_f32 = FlacArray.read_zarr(zf, mpi_comm=self.comm)
 
-        f64_file = os.path.join(tmppath, "data_f64.h5")
-        with H5File(f64_file, "w", comm=self.comm) as hf:
-            flcarr_f64.write_hdf5(hf.handle)
-        if self.comm is not None:
-            self.comm.barrier()
-        with H5File(f64_file, "r", comm=self.comm) as hf:
-            check_f64 = FlacArray.read_hdf5(hf.handle, mpi_comm=self.comm)
+        # f64_file = os.path.join(tmppath, "data_f64.zarr")
+        # with ZarrGroup(f64_file, mode="w", comm=self.comm) as zf:
+        #     flcarr_f64.write_zarr(zf)
+        # if self.comm is not None:
+        #     self.comm.barrier()
+        # with ZarrGroup(f64_file, mode="r", comm=self.comm) as zf:
+        #     check_f64 = FlacArray.read_zarr(zf, mpi_comm=self.comm)
 
-        del tmppath
-        del tmpdir
+        # if self.comm is not None:
+        #     self.comm.barrier()
+        # del tmppath
+        # del tmpdir
 
-        if check_i32 != flcarr_i32:
-            print(f"check_i32 = {check_i32}", flush=True)
-            print(f"flcarr_i32 = {flcarr_i32}", flush=True)
-            print("FAIL on i32 roundtrip to hdf5", flush=True)
-            self.assertTrue(False)
-        else:
-            output_i32 = check_i32.to_array(use_threads=True)
-            if not np.array_equal(output_i32, input32):
-                print("FAIL on i32 hdf5 decompressed array check", flush=True)
-                self.assertTrue(False)
+        # if check_i32 != flcarr_i32:
+        #     print(f"check_i32 = {check_i32}", flush=True)
+        #     print(f"flcarr_i32 = {flcarr_i32}", flush=True)
+        #     print("FAIL on i32 roundtrip to zarr", flush=True)
+        #     self.assertTrue(False)
+        # else:
+        #     output_i32 = check_i32.to_array()
+        #     if not np.array_equal(output_i32, input32):
+        #         print("FAIL on i32 zarr decompressed array check", flush=True)
+        #         self.assertTrue(False)
 
-        if check_i64 != flcarr_i64:
-            print(f"check_i64 = {check_i64}", flush=True)
-            print(f"flcarr_i64 = {flcarr_i64}", flush=True)
-            print("FAIL on i64 roundtrip to hdf5", flush=True)
-            self.assertTrue(False)
-        else:
-            output_i64 = check_i64.to_array(use_threads=True)
-            if not np.array_equal(output_i64, input64):
-                print("FAIL on i64 hdf5 decompressed array check", flush=True)
-                self.assertTrue(False)
+        # if check_i64 != flcarr_i64:
+        #     print("FAIL on i64 roundtrip to zarr", flush=True)
+        #     self.assertTrue(False)
+        # else:
+        #     output_i64 = check_i64.to_array()
+        #     if not np.array_equal(output_i64, input64):
+        #         print("FAIL on i64 zarr decompressed array check", flush=True)
+        #         self.assertTrue(False)
 
-        if check_f32 != flcarr_f32:
-            print("FAIL on f32 roundtrip to hdf5", flush=True)
-            self.assertTrue(False)
-        else:
-            output_f32 = check_f32.to_array(use_threads=True)
-            if not np.allclose(output_f32, inputf32, rtol=1.0e-5, atol=1.0e-5):
-                print("FAIL on f32 hdf5 decompressed array check", flush=True)
-                self.assertTrue(False)
+        # if check_f32 != flcarr_f32:
+        #     print("FAIL on f32 roundtrip to zarr", flush=True)
+        #     self.assertTrue(False)
+        # else:
+        #     output_f32 = check_f32.to_array()
+        #     if not np.allclose(output_f32, inputf32, rtol=1.0e-5, atol=1.0e-5):
+        #         print("FAIL on f32 zarr decompressed array check", flush=True)
+        #         self.assertTrue(False)
 
-        if check_f64 != flcarr_f64:
-            print("FAIL on f64 roundtrip to hdf5", flush=True)
-            self.assertTrue(False)
-        else:
-            output_f64 = check_f64.to_array(use_threads=True)
-            if not np.allclose(output_f64, inputf64, rtol=1.0e-5, atol=1.0e-5):
-                print("FAIL on f64 hdf5 decompressed array check", flush=True)
-                self.assertTrue(False)
+        # if check_f64 != flcarr_f64:
+        #     print("FAIL on f64 roundtrip to zarr", flush=True)
+        #     self.assertTrue(False)
+        # else:
+        #     output_f64 = check_f64.to_array()
+        #     if not np.allclose(output_f64, inputf64, rtol=1.0e-5, atol=1.0e-5):
+        #         print("FAIL on f64 zarr decompressed array check", flush=True)
+        #         self.assertTrue(False)
